refactor: #13 architecture
主流ファイルを変更
